refactor: 가시성 제어자 수정 및 패키지 이동
- SettingAppBar 파일 component 패키지 내부로 이동
- TuripAppBar Preview 관련 파일 가시성 제어자 private으로 수정
